+++ src/types/config.ts
@@ -47,6 +47,8 @@ export interface CHConfig extends DataSourceJsonData {
 
   httpHeaders?: CHHttpHeader[];
   forwardGrafanaHeaders?: boolean;
+  oauthPassThru?: boolean;
+  oauthPassThruAllowFallback?: boolean;
 
   customSettings?: CHCustomSetting[];
   enableSecureSocksProxy?: boolean;